# backend/pkg/restapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a REST API for PoeticMetric.",
        "title": "PoeticMetric REST API",
        "termsOfService": "https://poeticmetric.com/terms-of-service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "UserAccessTokenAuthentication": []
                    }
                ],
                "description": "List users.",
                "tags": [
                    "users"
                ],
                "summary": "List users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/poeticmetric.OrganizationUser"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserAccessTokenAuthentication": []
                    }
                ],
                "description": "Invite a new user to organization.",
                "tags": [
                    "users"
                ],
                "summary": "Invite user",
                "parameters": [
                    {
                        "description": "Params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/poeticmetric.InviteOrganizationUserParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/poeticmetric.OrganizationUser"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userID}": {
            "get": {
                "security": [
                    {
                        "UserAccessTokenAuthentication": []
                    }
                ],
                "description": "Read a user from the organization.",
                "tags": [
                    "users"
                ],
                "summary": "Read user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/poeticmetric.OrganizationUser"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserAccessTokenAuthentication": []
                    }
                ],
                "description": "Delete a user from the organization.",
                "tags": [
                    "users"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "UserAccessTokenAuthentication": []
                    }
                ],
                "description": "Update organization user.",
                "tags": [
                    "users"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/poeticmetric.UpdateOrganizationUserParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/poeticmetric.OrganizationUser"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "poeticmetric.InviteOrganizationUserParams": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "poeticmetric.OrganizationUser": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isEmailVerified": {
                    "type": "boolean"
                },
                "isOrganizationOwner": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "poeticmetric.UpdateOrganizationUserParams": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuthentication": {
            "type": "basic"
        },
        "UserAccessTokenAuthentication": {
            "description": "User access token authentication.",
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    }
}